+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -14,7 +14,7 @@ import seedu.address.logic.parser.AddressBookParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Contact;
 import seedu.address.storage.Storage;
 
 /**
