+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -18,8 +18,8 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Contact;
-import seedu.address.model.person.exceptions.DuplicatePersonException;
+import seedu.address.model.contact.Contact;
+import seedu.address.model.contact.exceptions.DuplicatePersonException;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddressBookTest {
