+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -16,7 +16,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.model.AddressBook;
-import seedu.address.model.person.Contact;
+import seedu.address.model.contact.Contact;
 
 /**
  * A utility class containing a list of {@code Contact} objects to be used in tests.
